# 02.backend_middleware/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智慧梁场中介服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/meta/enums": {
      "get": {
        "tags": [
          "contract"
        ],
        "summary": "Enums",
        "operationId": "enums_api_v1_meta_enums_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/projects/current": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project",
        "operationId": "project_api_v1_projects_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/yard-tree": {
      "get": {
        "tags": [
          "yard-areas"
        ],
        "summary": "Tree",
        "operationId": "tree_api_v1_yard_tree_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_YardAreaTreeNode__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/beams/by-code/{beam_code}/status": {
      "post": {
        "tags": [
          "beams"
        ],
        "summary": "Status",
        "operationId": "status_api_v1_beams_by_code__beam_code__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "beam_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beam Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeamStatusChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_BeamRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/beams/by-code/{beam_code}": {
      "get": {
        "tags": [
          "beams"
        ],
        "summary": "Beam",
        "operationId": "beam_api_v1_beams_by_code__beam_code__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "beam_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beam Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_BeamRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plc/telemetry": {
      "post": {
        "tags": [
          "plc"
        ],
        "summary": "Telemetry",
        "operationId": "telemetry_api_v1_plc_telemetry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/plc/process-records": {
      "post": {
        "tags": [
          "plc"
        ],
        "summary": "Process Report",
        "operationId": "process_report_api_v1_plc_process_records_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeamProcessExecutionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_BeamProcessExecutionRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ue5/contract": {
      "get": {
        "tags": [
          "ue5"
        ],
        "summary": "Contract",
        "operationId": "contract_api_v1_ue5_contract_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ue5/events": {
      "get": {
        "tags": [
          "ue5"
        ],
        "summary": "Events",
        "operationId": "events_api_v1_ue5_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_CursorPageResult_BeamLifecycleEventSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ue5/telemetry/latest": {
      "get": {
        "tags": [
          "ue5"
        ],
        "summary": "Latest",
        "operationId": "latest_api_v1_ue5_telemetry_latest_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "beam_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Beam Code"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-logs": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Audits",
        "operationId": "audits_api_v1_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageResult_OperationAuditLogSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/beam-positions": {
      "get": {
        "tags": [
          "beam-positions"
        ],
        "summary": "Listing",
        "operationId": "listing_api_v1_beam_positions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageResult_BeamPositionSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "beam-positions"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_beam_positions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeamPositionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_BeamPositionRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/beams": {
      "get": {
        "tags": [
          "beams"
        ],
        "summary": "Listing",
        "operationId": "listing_api_v1_beams_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageResult_BeamSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "beams"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_beams_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeamCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_BeamRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeamPositionCreate": {
        "properties": {
          "position_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Position Code"
          },
          "position_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Name"
          },
          "area_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Area Code"
          },
          "project_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Code"
          },
          "x_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,11}|(?=[\\d.]{1,15}0*$)\\d{0,11}\\.\\d{0,3}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Mm"
          },
          "y_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,11}|(?=[\\d.]{1,15}0*$)\\d{0,11}\\.\\d{0,3}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Mm"
          },
          "z_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,11}|(?=[\\d.]{1,15}0*$)\\d{0,11}\\.\\d{0,3}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Mm"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "position_code",
          "area_code"
        ],
        "title": "BeamPositionCreate"
      },
...
      "BeamProcessExecutionCreate": {
        "properties": {
          "execution_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Execution Code"
          },
          "project_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Code"
          },
          "beam_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Beam Code"
          },
          "process_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Process Code"
          },
          "result_code": {
            "$ref": "#/components/schemas/BeamProcessResult"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "actor_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Name"
          },
          "source": {
            "$ref": "#/components/schemas/RecordSource",
            "default": "MANUAL"
          },
          "external_record_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "External Record Id"
          },
          "supersedes_execution_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Execution Code"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "execution_code",
          "beam_code",
          "process_code",
          "result_code",
          "started_at",
          "finished_at"
        ],
        "title": "BeamProcessExecutionCreate"
      },
...
      "BeamProcessResult": {
        "type": "string",
        "enum": [
          "SUCCESS",
          "FAILED",
          "ABORTED"
        ],
        "title": "BeamProcessResult"
      },
...
      "BeamStatusChange": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BeamStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "BeamStatusChange"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Result_BeamPositionRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BeamPositionRead"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[BeamPositionRead]"
      },
...
      "Result_BeamProcessExecutionRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BeamProcessExecutionRead"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[BeamProcessExecutionRead]"
      },
...
      "Result_BeamRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BeamRead"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[BeamRead]"
      },
...
      "Result_CursorPageResult_BeamLifecycleEventSummary__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CursorPageResult_BeamLifecycleEventSummary_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[CursorPageResult[BeamLifecycleEventSummary]]"
      },
...
      "Result_PageResult_BeamPositionSummary__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageResult_BeamPositionSummary_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[PageResult[BeamPositionSummary]]"
      },
...
      "Result_PageResult_OperationAuditLogSummary__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PageResult_OperationAuditLogSummary_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[PageResult[OperationAuditLogSummary]]"
      },
...
      "Result_list_YardAreaTreeNode__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/YardAreaTreeNode"
            },
            "type": "array",
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "Result[list[YardAreaTreeNode]]"
      },
...
      "TelemetryReport": {
        "properties": {
          "message_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Message Id"
          },
          "device_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Device Id"
          },
          "project_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Code"
          },
          "beam_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Beam Code"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/Measurement"
            },
            "type": "array",
            "maxItems": 32,
            "minItems": 1,
            "title": "Measurements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_id",
          "device_id",
          "beam_code",
          "observed_at",
          "measurements"
        ],
        "title": "TelemetryReport"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}